refactor: move all text track helpers to utils/text-tracks.js
