+++ src/mse/virtual-source-buffer.js
@@ -2,9 +2,9 @@
  * @file virtual-source-buffer.js
  */
 import videojs from 'video.js';
-import createTextTracksIfNecessary from './create-text-tracks-if-necessary';
-import { removeCuesFromTrack } from './remove-cues-from-track';
-import { addTextTrackData } from './add-text-track-data';
+import { createTextTracksIfNecessary } from '../util/text-tracks.js';
+import { removeCuesFromTrack } from '../util/text-tracks.js';
+import { addTextTrackData } from '../util/text-tracks.js';
 import TransmuxWorker from 'worker!./transmuxer-worker.worker.js';
 import { isAudioCodec, isVideoCodec } from '../util/codecs';
 import {
